* commands/lsmmap.c (grub_cmd_lsmmap): Add casts to avoid printf 	warnings. 	* kern/ieee1275/openfw.c (grub_claimmap): Likewise. 	* disk/ieee1275/ofdisk.c (grub_ofdisk_open, grub_ofdisk_close, 	grub_ofdisk_read): Likewise, and deal similarly with the fact that 	ihandles have a 32-bit type but need to be stored in a "void *".

+++ commands/lsmmap.c
@@ -30,7 +30,7 @@ grub_cmd_lsmmap (grub_command_t cmd __attribute__ ((unused)),
   int NESTED_FUNC_ATTR hook (grub_uint64_t addr, grub_uint64_t size, grub_uint32_t type)
     {
       grub_printf ("base_addr = 0x%llx, length = 0x%llx, type = 0x%x\n",
-		   addr, size, type);
+		   (long long) addr, (long long) size, type);
       return 0;
     }
   grub_machine_mmap_iterate (hook);
